address issue within completed event handling
If multiple jobs are started at the same time it can happen that a runner
intended for one task picks up another one. During the shutdown phase
the wrong runner would be stopped and the workflow check fails. Instead
of a one to one relation between job run and runner it's a better option
to only spawn runners and to not decide which job they proceed. During
the complete event shutdown the runner based on the reported name (id).

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>