upgrade to linux kernel 6.1
This CL includes changes to the underlying bootstrapping of the micro vm.
Before this CL the NFS module was required within the kernel. As this
module provides a way to bootstrap the networking configuration. The
network configuration is available before the VM starts therefore it's a
good idea to pass the configuration into the VM instead of reaching out
to the metadataserver and relying on the NFS module to initialize the
rudimentary network stack first.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>