support waiting runners
When an GitHub Action Workflow has preconditions like for e.g. reviews
before beeing run it transitions into queued and afterswards into
waiting. You would assume the event to start a workflow run is only
issued after the preconditions are meth. But this is not the case GitHub
fires the events directly after the workflow is created. In order to
avoid launching runners for waiting jobs fetch the job instead of relying
on the event status.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>